CORE-39098 abstracted legacy optIn and Visitor functionality to modules
